+++ extensions/aiven/src/projects.tsx
@@ -1,21 +1,32 @@
 import { Action, ActionPanel, Color, Detail, getPreferenceValues, Icon, Keyboard, List } from "@raycast/api";
 import { useCachedState, useFetch } from "@raycast/utils";
-import { ErrorResult, ListProjectsResult, ListServiceBackupsResult, ListServicesResult, Project, Service, State } from "./interfaces";
+import {
+  ErrorResult,
+  ListProjectsResult,
+  ListServiceBackupsResult,
+  ListServiceLogsResult,
+  ListServicesResult,
+  Project,
+  Service,
+  State,
+} from "./interfaces";
 import { Fragment } from "react/jsx-runtime";
 import { filesize } from "filesize";
 
-function useAiven<T>(endpoint: string) {
+function useAiven<T>(endpoint: string, body?: Record<string, string>) {
   const { token } = getPreferenceValues<Preferences>();
   const { isLoading, data } = useFetch(`https://api.aiven.io/v1/${endpoint}`, {
+    method: !body ? "GET" : "POST",
     headers: {
       Authorization: `aivenv1 ${token}`,
       "Content-Type": "application/json",
     },
+    body: body ? JSON.stringify(body) : undefined,
     async parseResponse(response) {
       const result: ErrorResult | (object & T) = await response.json();
       if ("message" in result) throw new Error(result.message);
       return result;
-    }
+    },
   });
 
   return { isLoading, data };
@@ -45,12 +56,10 @@ export default function Projects() {
           }
           actions={
             <ActionPanel>
-              <Action.Push
-                icon={Icon.Coin}
-                title="View Services"
-                target={<Services project={project} />}
+              <Action.Push icon={Icon.Coin} title="View Services" target={<Services project={project} />} />
+              <Action.OpenInBrowser
+                url={`https://console.aiven.io/account/${project.account_id}/project/${project.project_name}/services`}
               />
-              <Action.OpenInBrowser url={`https://console.aiven.io/account/${project.account_id}/project/${project.project_name}/services`} />
             </ActionPanel>
           }
         />
@@ -91,9 +100,14 @@ ${Object.entries(service.metadata)
         return (
           <List.Item
             key={service.service_name}
-            icon={{ value: { source: Icon.CircleFilled, tintColor: getServiceIconColor(service.state) }, tooltip: service.state }}
+            icon={{
+              value: { source: Icon.CircleFilled, tintColor: getServiceIconColor(service.state) },
+              tooltip: service.state,
+            }}
             title={service.service_name}
-            subtitle={isShowingDetail ? undefined : { value: service.service_type, tooltip: service.service_type_description }}
+            subtitle={
+              isShowingDetail ? undefined : { value: service.service_type, tooltip: service.service_type_description }
+            }
             accessories={
               isShowingDetail
                 ? undefined
@@ -111,7 +125,13 @@ ${Object.entries(service.metadata)
                 metadata={
                   <List.Item.Detail.Metadata>
                     <List.Item.Detail.Metadata.TagList title="Features">
-                      {Object.entries(service.features).map(([feature, active]) => <List.Item.Detail.Metadata.TagList.Item key={feature} text={feature} color={active ? Color.Green : Color.Red} />)}
+                      {Object.entries(service.features).map(([feature, active]) => (
+                        <List.Item.Detail.Metadata.TagList.Item
+                          key={feature}
+                          text={feature}
+                          color={active ? Color.Green : Color.Red}
+                        />
+                      ))}
                     </List.Item.Detail.Metadata.TagList>
                     <List.Item.Detail.Metadata.Separator />
                     {service.components.map((component, index) => (
@@ -135,8 +155,21 @@ ${Object.entries(service.metadata)
                   title="Toggle Details"
                   onAction={() => setIsShowingDetail((show) => !show)}
                 />
-                <Action.Push icon={Icon.SaveDocument} title="View Backups" target={<Backups project={project} service={service} />} />
-                <Action.OpenInBrowser url={`https://console.aiven.io/account/${project.account_id}/project/${project.project_name}/services/${service.service_name}`} shortcut={Keyboard.Shortcut.Common.Open} />
+                <Action.Push
+                  icon={Icon.SaveDocument}
+                  title="View Backups"
+                  target={<Backups project={project} service={service} />}
+                />
+                <Action.Push
+                  icon={Icon.BulletPoints}
+                  title="View Logs"
+                  target={<Logs project={project} service={service} />}
+                  shortcut={{ modifiers: ["cmd"], key: "l" }}
+                />
+                <Action.OpenInBrowser
+                  url={`https://console.aiven.io/account/${project.account_id}/project/${project.project_name}/services/${service.service_name}`}
+                  shortcut={Keyboard.Shortcut.Common.Open}
+                />
               </ActionPanel>
             }
           />
@@ -146,30 +179,74 @@ ${Object.entries(service.metadata)
   );
 }
 
-function Backups({project, service}: {project: Project; service: Service}) {
-  const { isLoading, data } = useAiven<ListServiceBackupsResult>(`project/${project.project_name}/service/${service.service_name}/backups`);
+function Backups({ project, service }: { project: Project; service: Service }) {
+  const { isLoading, data } = useAiven<ListServiceBackupsResult>(
+    `project/${project.project_name}/service/${service.service_name}/backups`,
+  );
 
   function formatDate(dateString: string) {
     const date = new Date(dateString);
-    const options: Intl.DateTimeFormatOptions = { 
-      day: "2-digit", 
-      month: "short", 
-      year: "numeric", 
-      hour: "2-digit", 
-      minute: "2-digit", 
-      second: "2-digit", 
-      timeZone: "UTC", 
-      hour12: false 
-  };
-    const formattedDate = date.toLocaleString('en-US', options).replace(',', '');  
+    const options: Intl.DateTimeFormatOptions = {
+      day: "2-digit",
+      month: "short",
+      year: "numeric",
+      hour: "2-digit",
+      minute: "2-digit",
+      second: "2-digit",
+      timeZone: "UTC",
+      hour12: false,
+    };
+    const formattedDate = date.toLocaleString("en-US", options).replace(",", "");
     return formattedDate + " UTC";
   }
-  
-  return <Detail isLoading={isLoading} markdown={!data ? "" : `
+
+  return (
+    <Detail
+      isLoading={isLoading}
+      markdown={
+        !data
+          ? ""
+          : `
 | Data size | Backup time | Backup location |
 |-----------|-------------|-----------------|
-${data.backups.map(backup => `| ${filesize(backup.data_size, { base: 2 })} | ${formatDate(backup.backup_time)} | ${backup.storage_location} |`).join("\n")}
-`} actions={<ActionPanel>
-  <Action.OpenInBrowser url={`https://console.aiven.io/account/${project.account_id}/project/${project.project_name}/services/${service.service_name}/backups`} />
-</ActionPanel>} />
+${data.backups.map((backup) => `| ${filesize(backup.data_size, { base: 2 })} | ${formatDate(backup.backup_time)} | ${backup.storage_location} |`).join("\n")}
+`
+      }
+      actions={
+        <ActionPanel>
+          <Action.OpenInBrowser
+            url={`https://console.aiven.io/account/${project.account_id}/project/${project.project_name}/services/${service.service_name}/backups`}
+          />
+        </ActionPanel>
+      }
+    />
+  );
+}
+
+function Logs({ project, service }: { project: Project; service: Service }) {
+  const { isLoading, data } = useAiven<ListServiceLogsResult>(
+    `project/${project.project_name}/service/${service.service_name}/logs`,
+    {},
+  );
+
+  return (
+    <List isLoading={isLoading} isShowingDetail>
+      {data?.logs.map((log, index) => (
+        <List.Item
+          key={index}
+          icon={Icon.BulletPoints}
+          title={log.time}
+          detail={<List.Item.Detail markdown={log.msg} />}
+          actions={
+            <ActionPanel>
+              <Action.CopyToClipboard content={log.msg} />
+              <Action.OpenInBrowser
+                url={`https://console.aiven.io/account/${project.account_id}/project/${project.project_name}/services/${service.service_name}/logs`}
+              />
+            </ActionPanel>
+          }
+        />
+      ))}
+    </List>
+  );
 }