+++ extensions/aiven/CHANGELOG.md
@@ -1,9 +1,10 @@
 # Aiven Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Open in Browser + View Backups & Logs] - {PR_MERGE_DATE}
 
 - Add "Open in Browser" `Action`s
 - Add `tooltip`s to service items
 - View Service **Backups**
+- View Service **Logs**
 
 ## [Initial Version] - 2025-01-09